+++ retz-admin/src/main/java/io/github/retz/admin/MainCommand.java
@@ -22,7 +22,7 @@ import com.beust.jcommander.Parameter;
 
 public class MainCommand {
     @Parameter(names={"--config", "-C"})
-    private String configFile;
+    private String configFile = "/opt/retz-server/etc/retz.properties";
 
     @Parameter(names={"--verbose", "-v"})
     boolean verbose = false;